# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production-ready URL shortener API with JWT authentication, refresh-token rotation, and Google OAuth.\n\n## Authentication\nMost protected endpoints require an access token in the Authorization header:\n`Authorization: Bearer \u003caccess_token\u003e`\n\nAccess tokens are short-lived. Refresh tokens are stored in an HttpOnly cookie and used to obtain a new access token.\n\n## URL Expiration\nAuthenticated users receive longer URL expiration and cache durations.\nUnauthenticated URLs expire earlier.\n\n## Rate Limiting\nURL endpoints are rate limited to protect the service from abuse.",
        "title": "URL Shortener API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/urls": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all shortened URLs owned by the authenticated user.\n\nOnly URLs created by the current authenticated user are returned.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URLs"
                ],
                "summary": "Get current user's shortened URLs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_schema_url_schema.GetUrlsResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid access token",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a shortened URL.\n\nThe URL expiration depends on authentication status:\n- Authenticated users: longer expiration period.\n- Anonymous users: shorter expiration period.\n",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URLs"
                ],
                "summary": "Create a shortened URL",
                "parameters": [
                    {
                        "description": "URL to shorten",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_url.CreateUrlSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_schema_url_schema.ShortenUrlResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or URL",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/urls/{short_url}": {
            "get": {
                "description": "Resolves a shortened URL and redirects the client to the original destination.\n\nThe endpoint uses a cache-aside strategy:\n1. Redis cache is checked first.\n2. On cache miss, the URL is loaded from PostgreSQL.\n3. The result is cached with a TTL that never exceeds the URL expiration time.\n\nExpired URLs are treated as not found.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URLs"
                ],
                "summary": "Resolve a shortened URL",
                "parameters": [
                    {
                        "type": "string",
                        "example": "abc12345",
                        "description": "Short URL code",
                        "name": "short_url",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Temporary Redirect"
                    },
                    "400": {
                        "description": "Invalid short URL",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "URL not found or expired",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.Error"
                },
                "metaData": {
                    "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.MetaData"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.MetaData": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "github_com_tim8912097887-sys_url-shortener_internal_shared_schema_url_schema.GetUrlsResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "urls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_schema_url_schema.GetUrlsServiceResponse"
                    }
                }
            }
        },
        "github_com_tim8912097887-sys_url-shortener_internal_shared_schema_url_schema.GetUrlsServiceResponse": {
            "type": "object",
            "properties": {
                "expired_at": {
                    "type": "string"
                },
                "long_url": {
                    "type": "string"
                },
                "short_url": {
                    "type": "string"
                }
            }
        },
        "github_com_tim8912097887-sys_url-shortener_internal_shared_schema_url_schema.ShortenUrlResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "shortUrl": {
                    "type": "string"
                }
            }
        },
...
        "internal_url.CreateUrlSchema": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your access token in the format: Bearer {access_token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "RefreshCookie": {
            "description": "HttpOnly refresh token cookie automatically sent by the browser.",
            "type": "apiKey",
            "name": "refresh_token",
            "in": "cookie"
        }
    }
}